Add script to analyze the coverage of probes
This runs the hybseldesign's bin/analyze_probe_coverage.py
script across multiple datasets and over a provided
parameter space. It analyzes the coverage of probes given
in a FASTA file (e.g., the probes that were sequenced
following synthesis). It determines, for each dataset, the
best parameter set that yields a desired amount of coverage
over the genomes of the dataset.