Update script to generate E coli probe set with blacklisted sequences on more parameters
This generates probe sets for the following combinations of (mismatches, cover extension):
  (m, e) = (2, 25), (2, 50), (2, 100), (4, 25)
Although the script (run.sh) generates commands for more combinations, only the
above were actually run as of this commit.

The probe sets are in
  probe-sets/ecoli/v2-with-blacklist/uclust80/probes/
but are left out of this commit (they're large and not a final probe set).