Add plots summarizing V20 and V300 probe sets
This adds a script to make a stacked bar showing the number of
probes belonging to each dataset in a probe set, as well as a
script to make a bubble chart showing the number of datasets
that have each choice of parameters in a probe set. This also
adds the plots generated by the scripts for the V20 and V300
probe sets.

(Previously, I had generated these plots "manually" in macOS's
Numbers.)