Add tools to make plots that show how probe count scales with our approach vs baseline approaches
data/run.sh calls hybsel_design to count the number of probes
with a set cover approach and with various baseline approaches.
The baseline approaches are:
    * Tiling: tile across the genomes and get rid of exact duplicates
    * Dominating set (with various choices of mismatches ('m')): uses
      the naive dominating set approach to filter redundant probes,
      where redundancy is determined based on the length of the longest
      common substring up to m mismatches

The script make_plot.R produces a plot of how the number of probes
scales and takes data/[dataset]/num-probes.tsv as input.

This also includes data (number of probes, for various input sizes
and various approaches) for the 'ebola_zaire_with_2014' dataset.