Add data comparing number of probes on small input sizes
This adds probe counts for input sizes of n=1 and n=50 genomes.
It helps improve the look of the plots which use this data,
since now the lines (for the different approaches) converge
near n=1.

ebola_zaire_with_2014 already had data with an input size of
n=50, but this re-ran that.