Add table giving number of probes for parameter choices in V300 and V20
For V300 (all-human-viruses) and V20 (west-africa-viruses), this
adds a table that gives the number of probes required for each
combination of dataset and parameters (mismatches and cover
extension). These numbers are calculated from fasta files that
were generated during the design of these probe sets but are
ignored from the repo.