Fix download_dataset_fastas.py to be consistent in determing whether a dataset is segmented
It might be the case that sequences in a dataset had a segment name (e.g., "segment X") but
all sequence(s) have only one such name (i.e., all were labeled "segment X"). Then,
len(segments)=1 and this would determine that the datasets are segmented
when writing the fasta files but that they are not segmented when writing the `.py dataset
files.

See https://github.com/broadinstitute/catch/issues/17 for the resulting issues.

This changes the script to be consistent in determining whether a dataset is
segmented: as long as there is a segment name (even if just one), it determines
that the dataset is segmented.