Update viral accession list from GenBank using download from 2018-10
This includes an updated version of NCBI's viral accession list --
taxid10239.nbr.txt -- downloaded on 2018-10-19. It also updates
taxid10239.nbr.with_manual_additions.txt, which consists of the
new taxid10239.nbr.txt as well as all lines that were previously
manual additions but are not in the new taxid10239.nbr.txt.