Add option to combine all the .fasta files for segmented viruses into one
Previously, download_dataset_fastas.py wrote each genome (grouping
of sequences) for a segmented virus in its own fasta file, which
would create a lot of files for some segmented viruses. Now, it
still does the same, but the argument --consolidate-segmented-genomes-into-one-fasta
tells it to combine all of those fastas into one fasta file (and remove
the directory containing all of the previous ones).