Update viral sequence data using 2018-10-19 pull of accessions
See README.txt and ./run_from_lineages.sh for more detail on
how data was downloaded. This commit includes all accession
numbers in the download per-dataset (accession_nums/) and
(similarly) the seqs-in-201810-pull* files, but does not
include the data or the dataset .py files (these are in all/,
and are ignored by git).

This commit does not include any update to Influenza A/B/C data.