Update Influenza A/B/C sequence data using 2018-10-20 pull of accessions
This includes a new download of Influenza sequence accessions
(2018-10-20); see README.txt for details on this download. It
was converted into influenza.acc-list.txt (ignored from the
repo) using ./convert_to_acc_list.sh.

The viral sequence data was downloaded by running:
  ./download_influenza_seqs.sh
Its output is stored in download-genbank-viral-genomes/all/, which
is ignored from the repo. So the commit does not include the
data or the dataset .py files.