Add script to split Influenza A/B/C data into one FASTA per segment
This script takes a FASTA file containing all sequences for Influenza {A,B,C}
and creates new ones, each of which contains sequences for a segment.
The script also creates dataset .py files for each of the segments.

The input and output of the script is in download-genbank-viral-genomes/all/
so, although the script was run, its output is ignored from the
repo and not a part of this commit.