Fix bug in run.sh for Influenza probe design
This bug was introduced (and the script was committed) after the
script was actually run. It was run as described in README.txt.
As a result, the subcommand with the bug worked.

The call to `./run.sh combine-counts` (as made by the bug) is not
necessary. The subcommand is now named `param-exploration-combine-counts`
and, as described in README, it is run separately, before running
`./run.sh find-optimal-pooled-params`.