Remake respiratory probe set with fix in CATCH for `--expand-n`
Note that there are more probes than 50,000 in the 50k probe set, and
likewise for the 100k probe set. This is because the number of probes
for each parameter choice were selected a few days earlier; the designs
made for this commit incorporate more sequences for a few species (see
the additions in taxid-acc/), which increases the number of probes.