Remake V-Respiratory probe set (no substantive changes)
This commit contains no substantive changes in the probe set compared to
the previous commit; I had to regenerate only for a technical reason.

I mistakenly removed the dataset-runs/ data for hpiv_4 and
human_mastadenovirus_g (this data is not tracked by git). I removed this
because I thought I had supplied the wrong taxonomic ID for these
species. See the update in README.txt; in fact, the taxonomic IDs I had
for these were fine. Therefore, I had to redesign probes for these two
taxa to remake that data; there were no substantive changes in doing so,
but the probe counts change slightly due to stochasticity in design. I
then re-ran the process to pool, and the optimal parameters (in
pooled/design-100000/params.tsv) remained the same. I then remade the
pooled probe set, which is slightly different because of the small
differences in hpiv_4 and human_mastadenovirus_g.